+++ src/java/transform/Transformer.java
@@ -33,8 +33,8 @@ import org.eclipse.jdt.core.dom.rewrite.ListRewrite;
 import org.eclipse.jface.text.Document;
 import org.eclipse.text.edits.TextEdit;
 
-import sourceAnalysis.AnalyzedFile;
-import sourceAnalysis.AnalyzedMethod;
+import util.AnalyzedFile;
+import util.AnalyzedMethod;
 import transform.SymbolTable.SymbolTable;
 import transform.TypeChecking.TypeChecker;
 import transform.TypeChecking.TypeChecker.CType;
