+++ src/java/util/AnalyzedFile.java
@@ -1,4 +1,4 @@
-package sourceAnalysis;
+package util;
 
 import java.io.File;
 import java.util.HashSet;
