+++ src/test/unit/transformer/visitors/FinalizerVisitorTest.java
@@ -15,8 +15,8 @@ import org.eclipse.jface.text.Document;
 import org.eclipse.text.edits.TextEdit;
 import org.junit.Test;
 
-import sourceAnalysis.AnalyzedFile;
-import sourceAnalysis.AnalyzedMethod;
+import util.AnalyzedFile;
+import util.AnalyzedMethod;
 import transform.Transformer;
 import transform.SymbolTable.SymbolTable;
 import transform.TypeChecking.TypeChecker;
